pyproject.toml
[tool.poetry]
name = "effectivemobilelearning"
version = "0.1.0"
description = ""
authors = ["frank <frank.danil704@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-mock = "^3.14.0"
aioredis = "^2.0.1"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
